+++ src/living_world/cognition/__init__.py
@@ -5,7 +5,17 @@ from living_world.cognition.consolidation import (
     SleepCognitiveConsolidator,
 )
 from living_world.cognition.information_boundary import NPCInformationBoundary
+from living_world.cognition.llama_cpp_cognition_client import LlamaCppCognitionClient
+from living_world.cognition.npc_cognition_client import (
+    ActionOption,
+    ActionRequest,
+    NPCCognitionClient,
+    NPCCognitionClientError,
+    NPCCognitionInvalidResponseError,
+    NPCDecision,
+)
 from living_world.cognition.npc_context import NPCContext, NPCContextAssembler
+from living_world.cognition.ollama_cognition_client import OllamaCognitionClient
 from living_world.cognition.retrieval import (
     CognitiveRetriever,
     DeterministicCognitiveRetriever,
@@ -17,15 +27,23 @@ from living_world.managers.knowledge_manager import KnowledgeManager
 
 __all__ = [
     "DAY_LENGTH_TICKS",
+    "ActionOption",
+    "ActionRequest",
     "CognitiveConsolidationSystem",
     "CognitiveConsolidator",
     "CognitiveRetriever",
     "DeterministicCognitiveRetriever",
     "Knowledge",
     "KnowledgeManager",
+    "LlamaCppCognitionClient",
+    "NPCCognitionClient",
+    "NPCCognitionClientError",
+    "NPCCognitionInvalidResponseError",
     "NPCContext",
     "NPCContextAssembler",
+    "NPCDecision",
     "NPCInformationBoundary",
+    "OllamaCognitionClient",
     "RetrievalQuery",
     "RetrievedCognition",
     "SleepCognitiveConsolidator",